TEST: Spliting ArribaManagementServiceTests
Using different file for each service improve readability
