+++ src/Arriba/Arriba.Test/Services/GetTableInformationServiceTests.cs
@@ -1,0 +1,31 @@
+﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
+using System.Linq;
+
+namespace Arriba.Test.Services
+{
+    public partial class ArribaManagementServiceTests
+    {
+        [DataTestMethod]
+        [DataRow(TableName)]
+        public void GetTableInformationForUser(string tableName)
+        {
+            var table = _service.GetTableInformationForUser(tableName, _nonAuthenticatedUser);
+            Assert.IsNull(table);
+
+            var tableOwner = _service.GetTableInformationForUser(tableName, _owner);
+            Assert.IsNotNull(tableOwner);
+            Assert.IsTrue(tableOwner.CanAdminister);
+            Assert.IsTrue(tableOwner.CanWrite);
+
+            var tableWriter = _service.GetTableInformationForUser(tableName, _writer);
+            Assert.IsNotNull(tableWriter);
+            Assert.IsFalse(tableWriter.CanAdminister);
+            Assert.IsTrue(tableWriter.CanWrite);
+
+            var tableReader = _service.GetTableInformationForUser(tableName, _reader);
+            Assert.IsNotNull(tableReader);
+            Assert.IsFalse(tableReader.CanAdminister);
+            Assert.IsFalse(tableReader.CanWrite);
+        }
+    }
+}
